requirements.txt
Click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
pytest==6.2.2
WTForms == 3.0.1
flask-login == 0.6.2
flask-WTF == 1.0.1
setuptools==65.4.1
requests==2.21.0
Flask-Migrate==2.5.2
flask-mail==0.9.1
blinker==1.5
flask_bcrypt==1.0.1
bcrypt==4.0.1
pillow==9.2.0
dnspython==2.2.1 
email_validator==1.3.0
SQLAlchemy-serializer==1.4.1
furl==2.1.3
orderedmultidict==1.0.1
six==1.16.0